Reduce diff with Linux 5.14
Differences were reviewed using e.g.:

    diff -Naur -pX diffignore \
      drm-kmod/drivers/gpu/drm/i915/ \
      linux/drivers/gpu/drm/i915/

Where `diffignore` contains:

    Kconfig
    Kconfig.*
    Makefile

    drm_bridge_connector.h
    drm_fb_cma_helper.h
    drm_gem_atomic_helper.h
    drm_lease.h
    drm_managed.h
    drm_mipi_dbi.h
    drm_mipi_dsi.h
    drm_of.h
    bridge

    amdkfd
    amdgpu_atpx_handler.c
    amdgpu_trace_freebsd.h
    amdgpu_dm_trace_freebsd.h

    i915_trace.h
    i915_trace_points.c
    intel_acpi.c
    intel_freebsd.c
    selftests
    selftest_*.c
    selftest_*.h
    st_shmem_utils.c

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_bo_list.c
@@ -71,7 +71,7 @@ int amdgpu_bo_list_create(struct amdgpu_device *adev, struct drm_file *filp,
 	unsigned i;
 	int r;
 
-	if ((long)num_entries > (SIZE_MAX - sizeof(struct amdgpu_bo_list))
+	if (num_entries > (SIZE_MAX - sizeof(struct amdgpu_bo_list))
 				/ sizeof(struct amdgpu_bo_list_entry))
 		return -EINVAL;
